Added details to ClusterProvisioner overview
I've added more description and examples of the _ClusterProvisioner_ as I found the comments on Slack easier to understand than the current text. Given you've already edited the spec to change it to _ClusterProvisioner_ I thought it might make most sense to PR the update here rather than directly to PR482. I'm very happy to update this if I've misunderstood anything or it can be improved in other ways.